split TgSQL.cpp into TgSQL.cpp, BaseSQL.cpp and BaseSQL.h
